backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.7
pydantic-settings==2.0.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
neo4j==5.15.0
google-generativeai==0.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
networkx==3.2.1
aioredis==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
